implemented configmap reconcilation

+++ pkg/reconcilation/finalization/finalize.go
@@ -12,8 +12,21 @@ import (
 const Finalizer string = "echosec.jnnkrdb.de/finalizer"
 
 // finalize the given object
-func Finalize(ctx context.Context, c client.Client, obj client.Object) (bool, error) {
-	var _log = log.FromContext(ctx).WithValues("kind", obj.GetObjectKind().GroupVersionKind().Kind)
+func Finalize(ctx context.Context, c client.Client, obj client.Object, objType uint) (bool, error) {
+	var _log = log.FromContext(ctx).WithValues("objType", objType, "kind", obj.GetObjectKind().GroupVersionKind().Kind)
+
+	// ignore object, which are whether copy nor source
+	if objType == reconcilation.ObjectIsNONE {
+		// if the object already has a finalizer, then remove it
+		if controllerutil.ContainsFinalizer(obj, Finalizer) {
+			controllerutil.RemoveFinalizer(obj, Finalizer)
+			if err := c.Update(ctx, obj, &client.UpdateOptions{}); err != nil {
+				_log.Error(err, "error removing unneccessary finalizer")
+				return false, err
+			}
+		}
+		return false, nil
+	}
 
 	// checking the object for the finalizer, if the finalizer does not exist
 	// then append it to the object
@@ -33,14 +46,24 @@ func Finalize(ctx context.Context, c client.Client, obj client.Object) (bool, er
 		return false, nil
 	}
 
-	// create the delete options with the correct label selector
-	var deleteAllOptions = &client.DeleteAllOfOptions{}
-	deleteAllOptions.LabelSelector = reconcilation.ObjectsLabelSelector(obj.GetUID())
+	// if the object is a source object, all related objects
+	// must be removed from the cluster, if not, then only the
+	// finalizer will be removed from the copy object, so it can
+	// be terminated.
+	// If the copy object is still required in the cluster, and the#
+	// corresponding namespace exists, then it will be recreated in the
+	// next reconcilation run
+	if objType == reconcilation.ObjectIsSOURCE {
 
-	// remove all items from the cluster
-	if err := c.DeleteAllOf(ctx, obj, deleteAllOptions); err != nil {
-		_log.Error(err, "error removing all objects from cluster with specific labelselector", "labelselector", deleteAllOptions.LabelSelector)
-		return false, err
+		// create the delete options with the correct label selector
+		var deleteAllOptions = &client.DeleteAllOfOptions{}
+		deleteAllOptions.LabelSelector = reconcilation.ObjectsLabelSelector(obj.GetUID())
+
+		// remove all items from the cluster
+		if err := c.DeleteAllOf(ctx, obj, deleteAllOptions); err != nil {
+			_log.Error(err, "error removing all objects from cluster with specific labelselector", "labelselector", deleteAllOptions.LabelSelector)
+			return false, err
+		}
 	}
 
 	// after finalization remove the finalizer from the object
